+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+### Added:
+- Support reading a Parquet file into an RDD
+
 ## [0.1.3] - 2017-11-11
 ### Added:
 - Support minimal `SparkSession` class